purging SOIL, loading bmp images from scratch, still some issues

+++ mathHelper.cpp
@@ -9,10 +9,10 @@
 
 #include "mathHelper.h"
 
-/* 
+/*
  * translate
  *
- * INPUT: 
+ * INPUT:
  *         tx - translation on the x axis.
  *         ty - translation on the y axis.
  *         tz - translation on the z axis.
@@ -23,7 +23,7 @@
  * DESCRIPTION:
  *         This function is responsible for creating a
  *         a translation matrix given the inputs of the
- *         translation in each direction. 
+ *         translation in each direction.
  *
  */
 Matrix translate ( float tx, float ty, float tz ) {
@@ -36,10 +36,10 @@ Matrix translate ( float tx, float ty, float tz ) {
     return mTranslate;
 }
 
-/* 
+/*
  * scale
  *
- * INPUT: 
+ * INPUT:
  *         sx - scale on the x axis.
  *         sy - scale on the y axis.
  *         sz - scale on the z axis.
@@ -50,7 +50,7 @@ Matrix translate ( float tx, float ty, float tz ) {
  * DESCRIPTION:
  *         This function is responsible for creating a
  *         a scale matrix given the inputs of the
- *         scale in each direction. 
+ *         scale in each direction.
  *
  */
 Matrix scale ( float sx, float sy, float sz ) {
@@ -63,10 +63,10 @@ Matrix scale ( float sx, float sy, float sz ) {
     return mScale;
 }
 
-/* 
+/*
  * rotate
  *
- * INPUT: 
+ * INPUT:
  *         theta - angle (in degrees) of the rotation.
  *         vector - the angle the roation will use as a reference
  *
@@ -79,7 +79,7 @@ Matrix scale ( float sx, float sy, float sz ) {
  *         axis specified by the unity vector.
  *
  *         Reference:
- *         Belongie, Serge. "Rodrigues' Rotation Formula." 
+ *         Belongie, Serge. "Rodrigues' Rotation Formula."
  *         From MathWorld--A Wolfram Web Resource, created by Eric W. Weisstein.
  *         http://mathworld.wolfram.com/RodriguesRotationFormula.html
  *
@@ -91,20 +91,20 @@ Matrix rotate ( float theta, float vector[] ) {
     float wx = vector[0];
     float wy = vector[1];
     float wz = vector[2];
-    float rotate[] = { cosTheta + pow(wx,2) * (1 - cosTheta)     , wx * wy * (1 - cosTheta) - wz * sinTheta , wy * sinTheta + wx * wz * (1 - cosTheta)  , 0.0f ,
-                       wz * sinTheta + wx * wy * (1 - cosTheta)  , cosTheta + pow(wy,2) * (1 - cosTheta)    , -wx * sinTheta + wy * wz * (1 - cosTheta) , 0.0f ,
-                       -wy * sinTheta + wx * wz * (1 - cosTheta) , wx * sinTheta + wy * wz * (1 - cosTheta) , cosTheta + pow(wz,2) * (1 - cosTheta)     , 0.0f ,
+    float rotate[] = { static_cast<float>(cosTheta + pow(wx,2) * (1 - cosTheta))     , wx * wy * (1 - cosTheta) - wz * sinTheta , wy * sinTheta + wx * wz * (1 - cosTheta)  , 0.0f ,
+                       wz * sinTheta + wx * wy * (1 - cosTheta)  , static_cast<float>(cosTheta + pow(wy,2) * (1 - cosTheta))    , -wx * sinTheta + wy * wz * (1 - cosTheta) , 0.0f ,
+                       -wy * sinTheta + wx * wz * (1 - cosTheta) , wx * sinTheta + wy * wz * (1 - cosTheta) , static_cast<float>(cosTheta + pow(wz,2) * (1 - cosTheta))     , 0.0f ,
                                                              0.0f,                                      0.0f,                                       0.0f, 1.0f };
 
     Matrix mRotate(4,4,rotate);
 
-    return mRotate; 
+    return mRotate;
 }
 
-/* 
+/*
  * normalize
  *
- * INPUT: 
+ * INPUT:
  *         v - vector to be normalized.
  *
  * DESCRIPTION:
@@ -118,13 +118,13 @@ void normalize ( float v[] ) {
         v[0] = v[0] / len;
         v[1] = v[1] / len;
         v[2] = v[2] / len;
-    } 
+    }
 }
 
-/* 
+/*
  * dotProduct
  *
- * INPUT: 
+ * INPUT:
  *         n - first vector for the dot product.
  *         u - second vector for the dot product.
  *
